refactor: move app bar to separate component
